condensed extract v

+++ AFNI_BSM_ROI_BetaExtraction.csh
@@ -35,70 +35,84 @@ set task = (${study}_bsm)
 # III. INDIVIDUAL ANALYSES
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
 
-set ROI = IFG
-
-#dACC
-#L_dlPFC
-#R_dlPFC
-#IFG
-
 #set subjects = ($SUBJECT_LIST)
 #foreach SUBJECT ( `cat $subjects` )
 
 set subjects = hc001
 foreach SUBJECT ($subjects)
 
+foreach ROI (dACC L_dlPFC R_dlPFC IFG)
+
 setenv DATA_DIR ${SUBJECTS_DIR}/${SUBJECT}/${task}
 cd $DATA_DIR/results;
 
 echo "*******************************************************************************"
-echo " AFNI | BSM Analysis - Extract Betas | " ${SUBJECT}
+echo " AFNI | BSM Analysis - Extract $ROI Betas for $SUBJECT "
 echo "*******************************************************************************"
 
+
 echo "*******************************************************************************"
-echo " AFNI | Configure environment | " ${SUBJECT}
+echo " AFNI | 3dcopy & 3dresample | Fit masks to func data       | " 
+echo " AFNI | 3dbucket            | Stage func file to be masked | "
+echo " AFNI | 3dmaskave           | Average voxels in mask       | "
 echo "*******************************************************************************"
 
-echo "Removing previous $ROI files"
+if ($ROI == 'IFG') then
+
 rm *${ROI}*
 
-#echo "3dcopy $ROI mask"
-#3dcopy ${ROI_DIR}/${ROI}.nii ${ROI}.nii
-3dcopy ${ROI_DIR}/${ROI}+tlrc ${ROI}+tlrc
+3dcopy \
+${ROI_DIR}/${ROI}+tlrc ${ROI}+tlrc
 
-# Uncomment for IFG beta extraction
-#echo "creating IFG mask"
-#whereami -mask_atlas_region TT_Daemon::inf_frontal_gyrus
+3dresample \
+-master LSS.${SUBJECT}_despike+tlrc \
+-prefix ${ROI}_mask_resamp \
+-input ${ROI}+tlrc
 
-echo "*******************************************************************************"
-echo " AFNI | 3dbucket - Stage func file to be masked | " ${SUBJECT}
-echo "*******************************************************************************"
+3dbucket \
+-prefix ${ROI}_LSS_avg LSS.${SUBJECT}_despike+tlrc
 
-3dbucket -prefix ${ROI}_LSS_avg LSS.${SUBJECT}_despike+tlrc
-
-echo "*******************************************************************************"
-echo " AFNI | 3dresample - Force $ROI mask to match dimensions of BSM | " ${SUBJECT}
-echo "*******************************************************************************"
-
-#3dresample -master LSS.${SUBJECT}_despike+tlrc -prefix ${ROI}_mask_resamp -input ${ROI}.nii
-
-# Uncomment for IFG beta extraction
-3dresample -master LSS.${SUBJECT}_despike+tlrc -prefix ${ROI}_mask_resamp -input ${ROI}+tlrc
-
-echo "*******************************************************************************"
-echo " AFNI | 3dmaskave - Average voxels in $ROI mask | " ${SUBJECT}
-echo "*******************************************************************************"
-
-3dmaskave -quiet -mask ${ROI}_mask_resamp+tlrc ${ROI}_LSS_avg+tlrc > ${ROI}_LSS_avg_file.1D
+3dmaskave \
+-quiet \
+-mask ${ROI}_mask_resamp+tlrc ${ROI}_LSS_avg+tlrc > ${ROI}_LSS_avg_file.1D
 
 1dplot ${ROI}_LSS_avg_file.1D
+
+else if ($ROI == 'dACC' || $ROI == 'L_dlPFC' || $ROI == 'R_dlPFC') then
+
+rm *${ROI}*
+
+3dcopy \
+${ROI_DIR}/${ROI}.nii ${ROI}.nii
+
+3dresample \
+-master LSS.${SUBJECT}_despike+tlrc \
+-prefix ${ROI}_mask_resamp \
+-input ${ROI}.nii
+
+3dbucket \
+-prefix ${ROI}_LSS_avg LSS.${SUBJECT}_despike+tlrc
+
+3dmaskave \
+-quiet \
+-mask ${ROI}_mask_resamp+tlrc ${ROI}_LSS_avg+tlrc > ${ROI}_LSS_avg_file.1D
+
+1dplot ${ROI}_LSS_avg_file.1D
+
+else
+echo "Check ROIs"
+
+endif
 
 echo "*******************************************************************************"
 echo " AFNI | Beta Series Method - Beta Extraction COMPLETE | " ${SUBJECT}
 echo "*******************************************************************************"
 
-cd ${ANALYSIS_DIR}
+## End ROI loop
+end
 
 ## End subject loop
 end
+
+cd ${ANALYSIS_DIR}
